+++ docs/html/search/all_10.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['write_78',['write',['../classace__common_1_1PrintStrBase.html#a21d9667e7407a4818e94ab3232e35fb3',1,'ace_common::PrintStrBase::write(uint8_t c) override'],['../classace__common_1_1PrintStrBase.html#a9c49abf25f6a4d4a1ca3d7bd35140d10',1,'ace_common::PrintStrBase::write(const uint8_t *buf, size_t size) override']]]
+  ['write_86',['write',['../classace__common_1_1PrintStrBase.html#a21d9667e7407a4818e94ab3232e35fb3',1,'ace_common::PrintStrBase::write(uint8_t c) override'],['../classace__common_1_1PrintStrBase.html#a9c49abf25f6a4d4a1ca3d7bd35140d10',1,'ace_common::PrintStrBase::write(const uint8_t *buf, size_t size) override']]]
 ];