+++ docs/html/search/all_e.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['url_20encoding_73',['Url Encoding',['../md__home_brian_src_AceCommon_src_url_encoding_README.html',1,'']]],
-  ['udiv1000_74',['udiv1000',['../arithmetic_8h.html#a239844f52c0a01e7fcd0075739035783',1,'ace_common']]],
-  ['update_75',['update',['../classace__common_1_1TimingStats.html#a7e2074483d3bace1f9695c573dac811e',1,'ace_common::TimingStats']]],
-  ['url_5fencoding_2eh_76',['url_encoding.h',['../url__encoding_8h.html',1,'']]]
+  ['url_20encoding_81',['Url Encoding',['../md__home_brian_src_AceCommon_src_url_encoding_README.html',1,'']]],
+  ['udiv1000_82',['udiv1000',['../arithmetic_8h.html#a239844f52c0a01e7fcd0075739035783',1,'ace_common']]],
+  ['update_83',['update',['../classace__common_1_1GenericStats.html#a6a8d12163d4b1977c86bb3c1ce17c16f',1,'ace_common::GenericStats::update()'],['../classace__common_1_1TimingStats.html#a7e2074483d3bace1f9695c573dac811e',1,'ace_common::TimingStats::update()']]],
+  ['url_5fencoding_2eh_84',['url_encoding.h',['../url__encoding_8h.html',1,'']]]
 ];