+++ docs/html/search/classes_0.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['fcstring_80',['FCString',['../classace__common_1_1FCString.html',1,'ace_common']]],
-  ['flashstring_81',['FlashString',['../classace__common_1_1FlashString.html',1,'ace_common']]]
+  ['fcstring_88',['FCString',['../classace__common_1_1FCString.html',1,'ace_common']]],
+  ['flashstring_89',['FlashString',['../classace__common_1_1FlashString.html',1,'ace_common']]]
 ];